% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{assemble_basis_metric}
\alias{assemble_basis_metric}
\title{assemble_basis_metric}
\usage{
assemble_basis_metric(basis_list, curves_to_keep = NULL)
}
\arguments{
\item{basis_list}{a list of basis fd object}

\item{curves_to_keep}{a list of the states curves to keep}
}
\value{
a matrix of the metric to consider
}
\description{
This function assemble the metrics of all the basis of the basis list.
This function only assemble the needed basis, especially if length(curve_to_keep) != N_states
}
\examples{
basis1 = fda::create.bspline.basis(c(0,100), nbasis=10, norder=4)
basis2 = fda::create.bspline.basis(c(0,100), nbasis=15, norder=1)
basis3 = fda::create.fourier.basis(c(0,100), nbasis=7)
assemble_basis_metric(list(basis1, basis2, basis3), list(1,2,4))
assemble_basis_metric(list(basis1, basis2, basis3), list(1,2))

}
\author{
Francois Bassac
}
