% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpls.R
\name{assert_funcPLS_inputs}
\alias{assert_funcPLS_inputs}
\title{assert_funcPLS_inputs}
\usage{
assert_funcPLS_inputs(
  df_list,
  Y,
  basis_obj,
  regul_time_obj,
  curve_type_obj = NULL,
  id_col_obj = "id",
  time_col_obj = "time"
)
}
\arguments{
\item{df_list}{a list of dataframes (id, time, value_or_state)}

\item{Y}{a numeric vector of the response}

\item{basis_obj}{a list of basis object or a basis object}

\item{regul_time_obj}{a vector of time regularization values or a list of vectors}

\item{curve_type_obj}{a character "cat" or 'num' or a list of those values}

\item{id_col_obj}{a character of the id column for all the curves or a list of id column character}

\item{time_col_obj}{a character of the time column for all the curves or a list of time column character}
}
\value{
a list of (basis_list, regul_time_list, curve_type_list,
id_col_list, time_col_list)
}
\description{
This function checks the integrity of the input for funcPLS.
It returns a list of (basis_list, regul_time_list, curve_type_list,
id_col_list, time_col_list)
}
\author{
Francois Bassac
}
