% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{assert_mpfr_inputs}
\alias{assert_mpfr_inputs}
\title{Check inputs for Multivariate Penalized Functional Regression}
\usage{
assert_mpfr_inputs(data_list, Y, types, basis_list)
}
\arguments{
\item{data_list}{A list of raw data objects (matrices for NFD, data.frames/fd for SFD/CFD).}

\item{Y}{A numeric vector of the response.}

\item{types}{A character vector specifying the type of each predictor ('nfd', 'sfd', 'cfd').}

\item{basis_list}{A list of fda basis objects. Can contain NULL for NFDs.}
}
\value{
TRUE if all checks pass, otherwise stops with an error.
}
\description{
Check inputs for Multivariate Penalized Functional Regression
}
