% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_pls.R
\name{assert_multivariate_naivePLS_inputs}
\alias{assert_multivariate_naivePLS_inputs}
\title{assert_multivariate_naivePLS_inputs}
\usage{
assert_multivariate_naivePLS_inputs(
  df_list,
  Y,
  regul_time_obj = NULL,
  curve_type_obj = NULL,
  id_col_obj = "id",
  time_col_obj = "time"
)
}
\arguments{
\item{df_list}{a list of dataframe (id, time, value_or_state)}

\item{Y}{a numeric vector}

\item{regul_time_obj}{a list of time regularisation values}

\item{curve_type_obj}{a list of curve type 'cat' or 'num'}

\item{id_col_obj}{a list of the names of the id columns}

\item{time_col_obj}{a list of the names of the time columns}
}
\value{
a list
}
\description{
This function checks the input of naivePLS function.
}
\author{
Francois Bassac
}
