% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{bdiag_base}
\alias{bdiag_base}
\title{Helper function: Create a block-diagonal matrix from a list of matrices
(Pure base R for speed and zero dependencies)}
\usage{
bdiag_base(mat_list)
}
\description{
Helper function: Create a block-diagonal matrix from a list of matrices
(Pure base R for speed and zero dependencies)
}
