% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{beta_1_real_func}
\alias{beta_1_real_func}
\title{beta_1_real_func}
\usage{
beta_1_real_func(t, end_time = 100, drop = NULL)
}
\arguments{
\item{t}{evaluation time}

\item{end_time}{end time; default 100}

\item{drop}{particular point of the curve, default NULL}
}
\value{
a value
}
\description{
beta_1_real_func
}
\examples{
beta_1_real_func(0)
beta_1_real_func(10)
beta_1_real_func(10:90)
plot(x=0:100, y=beta_1_real_func(0:100, 100), type='l', main="Beta_1")

}
\author{
Francois Bassac
}
