% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{beta_2_real_func}
\alias{beta_2_real_func}
\title{beta_2_real_func}
\usage{
beta_2_real_func(t, end_time = 100, drop = 3 * 100/5)
}
\arguments{
\item{t}{evaluation time}

\item{end_time}{end time; default 100}

\item{drop}{particular point of the curve, default 3*100/5}
}
\value{
a value
}
\description{
beta_2_real_func
}
\examples{
beta_2_real_func(0)
beta_2_real_func(10)
beta_2_real_func(10:90)
plot(x=0:100, y=beta_2_real_func(0:100, 100), type='l', main="Beta_2")

}
\author{
Francois Bassac
}
