% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{beta_6_real_func}
\alias{beta_6_real_func}
\title{beta_6_real_func}
\usage{
beta_6_real_func(t, end_time = 100, drop = 1)
}
\arguments{
\item{t}{evaluation time}

\item{end_time}{end time; default 100}

\item{drop}{particular point of the curve, default 1}
}
\value{
a value
}
\description{
Constant function = 1
Can adjuste the constant value by drop input
}
\examples{
beta_5_real_func(0)
beta_5_real_func(10)
beta_5_real_func(10:90)
plot(x=0:100, y=beta_5_real_func(0:100, 100), type='l', main="Beta_5")

}
\author{
Francois Bassac
}
