% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{beta_list_generation}
\alias{beta_list_generation}
\title{beta_list_generation}
\usage{
beta_list_generation(N_states = 3)
}
\arguments{
\item{N_states}{a int of the number of states wanted, default 3}
}
\value{
a list of functions
}
\description{
beta_list_generation
}
\examples{

beta_list = beta_list_generation()
beta_list_2 = beta_list_generation(6)

}
\author{
Francois Bassac
}
