% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{block_diag}
\alias{block_diag}
\title{block_diag}
\usage{
block_diag(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}
}
\value{
a matrix
}
\description{
returns a matrix whose blocks are A and B.
returns :
( A, 0 )
( 0, B )
}
\examples{
A = matrix(c(1, 2, 3, 4), 2)
B = matrix(c(5, 6, 7, 8,9, 10), 2)
C = block_diag(A, B)

}
\author{
Francois Bassac
}
