% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{build_block_cfd}
\alias{build_block_cfd}
\title{Build block for Categorical Functional Data (CFD)}
\usage{
build_block_cfd(df_cfd, basis_obj, reference_state = NULL, LDO = 2, ...)
}
\arguments{
\item{df_cfd}{The dataframe containing the CFD for all individuals}

\item{basis_obj}{The fda basis object}

\item{reference_state}{Character. The state to drop (K-1). If NULL, keeps all K states.}

\item{...}{Additional arguments to pass to your evaluate_lambda function}
}
\value{
A list with the concatenated design block Z and the base penalty R
}
\description{
Build block for Categorical Functional Data (CFD)
}
