% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{build_block_nfd}
\alias{build_block_nfd}
\title{Build block for Non-Functional Data (NFD / Scalars)}
\usage{
build_block_nfd(X_nfd)
}
\arguments{
\item{X_nfd}{A numeric matrix or data.frame of scalar predictors (n x Q)}
}
\value{
A list with the design block Z and the penalty base R (Identity)
}
\description{
Build block for Non-Functional Data (NFD / Scalars)
}
