% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{build_block_sfd}
\alias{build_block_sfd}
\title{Build block for Scalar Functional Data (SFD)}
\usage{
build_block_sfd(coef_matrix, basis_obj, LDO = 2)
}
\arguments{
\item{coef_matrix}{Matrix of coefficients (n x d) of the curves on their basis}

\item{basis_obj}{The fda basis object}

\item{LDO}{Linear Differential Operator for the penalty (default 2)}
}
\value{
A list with the design block Z and the penalty base R
}
\description{
Build block for Scalar Functional Data (SFD)
}
