% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{build_global_design}
\alias{build_global_design}
\title{Build the Global Design Matrix (D or Z)}
\usage{
build_global_design(block_list, n_obs)
}
\arguments{
\item{block_list}{A list of blocks generated by the build_block_* functions.}

\item{n_obs}{The number of observations (rows).}
}
\value{
A matrix of dimension n x (1 + P_total)
}
\description{
Concatenates the intercept and all individual predictor blocks horizontally.
}
