% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{build_global_penalty}
\alias{build_global_penalty}
\title{Build the Augmented Global Penalty Matrix (R_0_lambda)}
\usage{
build_global_penalty(block_list, lambda_list)
}
\arguments{
\item{block_list}{A list of blocks generated by the build_block_* functions.}

\item{lambda_list}{A list of numeric vectors containing the lambdas for each block.}
}
\value{
A symmetric square matrix representing R_0_lambda
}
\description{
Constructs the block-diagonal penalty matrix dynamically. Supports both
global lambda per predictor and specific lambdas per CFD state.
}
