% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{build_new_data_list}
\alias{build_new_data_list}
\title{build_new_data_list}
\usage{
build_new_data_list(
  df_list,
  N_curves,
  orth_basis_list = NULL,
  basis_list = NULL,
  curve_type_list,
  id_col_list,
  time_col_list,
  regul_time_list
)
}
\arguments{
\item{df_list}{a list of dataframes (id, time, value_or_state)}

\item{N_curves}{a integer, the number of curves}

\item{orth_basis_list}{a list of orthogonalized basis fd list}

\item{basis_list}{a list of basis fd functions}

\item{curve_type_list}{a list of the curve type of each curve}

\item{id_col_list}{a list of the id column name for each curve}

\item{time_col_list}{a list of the time column name for each curve}

\item{regul_time_list}{a list of the time regularization vector for each
curve}
}
\value{
a list
}
\description{
This function preprocess the different input in order to format them to the
right number of curve. Warning the "right" number of curve take into account
the number of different states for the CFDs.
}
\author{
Francois Bassac
}
