% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{build_reg_curve_spls}
\alias{build_reg_curve_spls}
\title{build_reg_curve_spls}
\usage{
build_reg_curve_spls(
  plsr_model,
  curves_names_list,
  v_i_list,
  nb_comp_pls_opt = NULL,
  print_steps = TRUE
)
}
\arguments{
\item{plsr_model}{a pls model}

\item{curves_names_list}{a list of the names of the different curves}

\item{v_i_list}{a list of the v functions}

\item{nb_comp_pls_opt}{a integer, the number of component to take into
account. if null (default) all the components are used}

\item{print_steps}{a boolean to print steps}
}
\value{
a list of fd object
}
\description{
This function builds the smooth PLS regression functions.
}
\author{
Francois Bassac
}
