% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{build_spls_functions}
\alias{build_spls_functions}
\title{build_spls_functions}
\usage{
build_spls_functions(
  curves_names_list,
  new_basis_list,
  new_orth_basis_list,
  d_i,
  u_i
)
}
\arguments{
\item{curves_names_list}{a list of the names of the different curves}

\item{new_basis_list}{a list of the initial basis fd object for all curves}

\item{new_orth_basis_list}{a list of the orthogonalized basis as fd list for
all curves}

\item{d_i}{the pls coefficient such as X = d_i t_i :
plsr_model$loadings}

\item{u_i}{the pls coefficient such as t_i = X u_i :
plsr_model$loading.weights}
}
\value{
a list Lambda = sum d_i t_i
}
\description{
This function build some intermediate functions of the smooth pls algorithm.
It also evaluates the different coefficients gamma_ij
}
\author{
Francois Bassac
}
