% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{build_u_ki_list}
\alias{build_u_ki_list}
\title{build_u_ki_list}
\usage{
build_u_ki_list(N_states, nbComp, ms_pls_models)
}
\arguments{
\item{N_states}{a integer, number of different states}

\item{nbComp}{a integer, max number of components}

\item{ms_pls_models}{a list of the intermediate pls models to evaluate the
real multi-stats pls components.}
}
\value{
a list of the u_i^k
}
\description{
build_u_ki_list
}
\author{
Francois Bassac
}
