% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{calc_penalty_matrix}
\alias{calc_penalty_matrix}
\title{Compute the penalty matrix for a functional basis}
\usage{
calc_penalty_matrix(basis_obj, LDO = 2)
}
\arguments{
\item{basis_obj}{A functional data object (\code{fd}), a basis object (\code{basis.fd}),
or a list of \code{fd} objects representing individual basis functions.}

\item{LDO}{A numeric integer defining the Linear Differential Operator (LDO).
Defaults to 2 for the second derivative penalty.}
}
\value{
A square symmetric numeric matrix representing the penalty matrix.
}
\description{
Calculates the penalty matrix associated with a specific Linear Differential
Operator (LDO). By default, it computes the inner product of the second
derivatives of the basis functions, which is typically used to enforce
smoothness in penalized functional regression.
}
\examples{
\dontrun{
# Assuming 'my_basis' is a basisfd object
pen_mat <- calc_penalty_matrix(my_basis, LDO = 2)
}
}
\author{
Francois Bassac
}
