% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{convert_to_wide_format}
\alias{convert_to_wide_format}
\title{convert_to_wide_format}
\usage{
convert_to_wide_format(df_new, id_col = "id", time_col = "time")
}
\arguments{
\item{df_new}{a regularized dataframe}

\item{id_col}{col_name of df_new for the id}

\item{time_col}{col_name of df_new for the id}
}
\value{
the dataframe in wide format
}
\description{
This function takes the df_new output from regularize_time_series and
convert it into another format
}
\examples{
id_df = data.frame(id=rep(1,5), time=seq(0, 40, 10), state=c(0, 1, 1, 0, 1))
id_df_new = regularize_time_series(id_df, time_seq = seq(0, 40, 2), curve_type = 'cat')
convert_to_wide_format(id_df_new)

}
\author{
Francois Bassac
}
