% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{create_bspline_basis}
\alias{create_bspline_basis}
\title{create_bspline_basis}
\usage{
create_bspline_basis(start, end, nbasis = 10, norder = 4)
}
\arguments{
\item{start}{start time}

\item{end}{end time}

\item{nbasis}{number of basis functions, default 10}

\item{norder}{order of the basis function, default cubic splines 4}
}
\value{
a basis fd object
}
\description{
create_bspline_basis
}
\examples{
b0 = create_bspline_basis(0, 10, 10, 4)
plot(b0)

b1 = create_bspline_basis(0, 10, 10, 2)
plot(b1)

b2 = create_bspline_basis(0, 10, 10, 1)
plot(b1)


}
\author{
Francois Bassac
}
