% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{cv_pfr_multi}
\alias{cv_pfr_multi}
\title{Multivariate Exact LOOCV for PFR (Parallelized)}
\usage{
cv_pfr_multi(block_list, Y, list_of_lambda_lists, parallel = FALSE)
}
\arguments{
\item{block_list}{A list of blocks generated by the build_block_* functions.}

\item{Y}{The response vector.}

\item{list_of_lambda_lists}{A list containing all the lambda combinations to test.}

\item{parallel}{Logical. If TRUE, executes the grid search in parallel.}
}
\value{
A list with the optimal lambda combination and the minimum RMSEP.
}
\description{
Multivariate Exact LOOCV for PFR (Parallelized)
}
