% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{cv_pfr_multi_no_parallel}
\alias{cv_pfr_multi_no_parallel}
\title{Multivariate Exact LOOCV for PFR}
\usage{
cv_pfr_multi_no_parallel(block_list, Y, list_of_lambda_lists)
}
\arguments{
\item{block_list}{A list of blocks generated by the build_block_* functions}

\item{Y}{The response vector}

\item{list_of_lambda_lists}{A list containing all the lambda combinations to test}
}
\value{
A list with the optimal lambda combination and the minimum RMSEP
}
\description{
Multivariate Exact LOOCV for PFR
}
