% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpls.R
\name{determine_curve_name}
\alias{determine_curve_name}
\title{determine_curve_name}
\usage{
determine_curve_name(curve_number, curve_type = NULL, states_names = NULL)
}
\arguments{
\item{curve_number}{a int, the place of the vurve in the data}

\item{curve_type}{a character, 'cat' or 'num'}

\item{states_names}{a character or list of character with the states of the CFD}
}
\value{
a vector of names.
}
\description{
This function determines the curves names bases on its place in the data and
its states if it is a categorical functional data.
}
\author{
Francois Bassac
}
