% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{determine_next_state}
\alias{determine_next_state}
\title{determine_next_state}
\usage{
determine_next_state(current_state, transition_df)
}
\arguments{
\item{current_state}{a value of the current state}

\item{transition_df}{a dataframe of the transistion matrix}
}
\value{
a value for the next state
}
\description{
This function determines the next state base on the current state and
the transition matrix.
}
\examples{
N_states = 3
lambdas = lambda_determination(N_states)
transition_df = transfert_probabilities(N_states)
determine_next_state(1, transition_df)

}
\author{
Francois Bassac
}
