% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{eval_max_min_y}
\alias{eval_max_min_y}
\title{eval_max_min_y}
\usage{
eval_max_min_y(f_list, regul_time)
}
\arguments{
\item{f_list}{a list of functions and fd objects}

\item{regul_time}{a vector of time evaluation points.}
}
\value{
a vector
}
\description{
This function returns the min and max values of a list of functions and fd
objects on regul_time.
}
\author{
Francois Bassac
}
