% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{evaluate_curves_distances}
\alias{evaluate_curves_distances}
\title{evaluate_curves_distances}
\usage{
evaluate_curves_distances(real_f, regul_time, fun_fd_list = NULL)
}
\arguments{
\item{real_f}{a fun of fd function, base function to compare}

\item{regul_time}{a vector of time regularization values}

\item{fun_fd_list}{a list of fun or fd functions or a fun or a fd function}
}
\value{
No return value, called for side effects (prints distances to the console).
}
\description{
Either for R func or fd function, this function evaluates the distance
between the real curve and each curve fun or fd which are in the
func_fd_list.
}
\author{
Francois Bassac
}
