% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{evaluate_gamma_ij}
\alias{evaluate_gamma_ij}
\title{evaluate_gamma_ij}
\usage{
evaluate_gamma_ij(w_i_list, p_i_list)
}
\arguments{
\item{w_i_list}{a list of fd functions w_i(t)}

\item{p_i_list}{a list of fd functions p_i(t)}
}
\value{
a list of list of the gamma_ij values
}
\description{
This function evaluates int_0^T w_i_fd p_j_t dt for all j
return a list, the element i length is (i-1)
}
\author{
Francois Bassac
}
