% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{evaluate_id_func_integral}
\alias{evaluate_id_func_integral}
\title{evaluate_id_func_integral}
\usage{
evaluate_id_func_integral(
  id_df,
  func,
  id_col = "id",
  time_col = "time",
  rel_tol = .Machine$double.eps^0.5,
  subdivisions = 1000L,
  ...
)
}
\arguments{
\item{id_df}{Dataframe for a single individual with at least columns (id, time, state).}

\item{func}{The R function to integrate.}

\item{id_col}{Character, name of the id column, default 'id'.}

\item{time_col}{Character, name of the time column, default 'time'.}

\item{rel_tol}{Relative tolerance for stats::integrate, default 1e-8.}

\item{subdivisions}{Max number of subdivisions for integrate, default 100.}

\item{...}{Additional arguments (ignored to prevent passing unused params to func).}
}
\value{
A dataframe with the id and the calculated integral value.
}
\description{
Evaluates the integral \eqn{\int_{\tau_i} f(t) dt} where \eqn{\tau_i} are the active
intervals of a categorical functional data (states 0 or 1).
}
