% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{evaluate_id_func_integral_deprecated}
\alias{evaluate_id_func_integral_deprecated}
\title{evaluate_id_func_integral}
\usage{
evaluate_id_func_integral_deprecated(
  id_df,
  func,
  mode = 1,
  id_col = "id",
  time_col = "time",
  nb_pt = 10,
  subdivisions = 100
)
}
\arguments{
\item{id_df}{a single id dataframe of at least named columns (id, time)}

\item{func}{the function to integrate}

\item{mode}{select the integration mode 1 for R function integrate,
2 for pracma::trapz. default value : 1}

\item{id_col}{col_name of df for the id}

\item{time_col}{col_name of df for the time}

\item{nb_pt}{number of points for the integration, default value : 10}

\item{subdivisions}{default parameter of R function integrate;
default value : 100}
}
\value{
a dataframe with the id and the integral value.
}
\description{
This function evaluate the integral for a state (0, 1) functional data :
int( X(t) func(t) )dt.
This function works ONLY for a one state CFD!
}
\examples{
id_df = data.frame(id=rep(1,5), time=seq(0, 40, 10), state=c(0, 1, 1, 0, 1))
evaluate_id_func_integral(id_df, function(t){t})

}
\author{
Francois Bassac
}
