% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{evaluate_lambda_CFD_para_v1}
\alias{evaluate_lambda_CFD_para_v1}
\title{evaluate_lambda_CFD_para_v1}
\usage{
evaluate_lambda_CFD_para_v1(
  df,
  basis,
  int_mode = 1,
  id_col = "id",
  time_col = "time",
  nb_pt = 10,
  subdivisions = 100,
  parallel = TRUE
)
}
\arguments{
\item{df}{dataframe X(t)}

\item{basis}{basis fd object or a list of fd functions}

\item{int_mode}{integration mode, 1 for integrate, 2 for pracma::trapz}

\item{id_col}{a character for the id column, default 'id'}

\item{time_col}{a character for the time column, default 'time'}

\item{nb_pt}{number of points for the integration, default value : 10}

\item{subdivisions}{default parameter of R function integrate;
default value : 100}

\item{parallel}{a boolean to enable parallel processing, default TRUE}
}
\value{
a matrix of dimension nbasis columns and nind rows
}
\description{
This function evaluates the Lambda matrix
Lambda_ij = int X_j(t) phi_i(t) dt using parallel processing.
}
\author{
Francois Bassac
}
