% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{evaluate_lambda_SFD}
\alias{evaluate_lambda_SFD}
\title{evaluate_lambda_SFD}
\usage{
evaluate_lambda_SFD(
  df,
  basis,
  regul_time = seq(basis$rangeval[1], basis$rangeval[2], 1),
  int_mode = 1,
  id_col = "id",
  time_col = "time",
  subdivisions = 100
)
}
\arguments{
\item{df}{dataframe X(t)}

\item{basis}{basis fd object}

\item{regul_time}{a vector of time regularization values default basis
rangeval per 1}

\item{int_mode}{int, integration mode, 1 for integrate, 2 for pracma::trapz}

\item{id_col}{default name of the id column}

\item{time_col}{default name of the time column}

\item{subdivisions}{default parameter of R function integrate;
default value : 100}
}
\value{
a matrix of dimension nbasis columns and nind rows
}
\description{
This function evaluates the Lambda matrix
Lambda_ij = int X_j(t) phi_i(t) dt for step > 1
}
\author{
Francois Bassac
}
