% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{evaluate_metric}
\alias{evaluate_metric}
\title{evaluate_metric}
\usage{
evaluate_metric(basis)
}
\arguments{
\item{basis}{basis to evaluate the metric}
}
\value{
a matrix of dimension nbasis X nbasis
}
\description{
This function evaluates the metric of a certain basis.
The metric is the inprod of the basis functions.
}
\examples{
basis = create_bspline_basis(start=0, end=10, nbasis=10, norder=4)
metric = evaluate_metric(basis)

basis1 = create_bspline_basis(start=0, end=20, nbasis=10, norder=1)
metric1 = evaluate_metric(basis1)

}
\author{
Francois Bassac
}
