% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{evaluate_reg_curve_SPLS_uni}
\alias{evaluate_reg_curve_SPLS_uni}
\title{evaluate_reg_curve_SPLS_uni}
\usage{
evaluate_reg_curve_SPLS_uni(plsr_model, v_i_list, nb_comp = NULL)
}
\arguments{
\item{plsr_model}{a model from pls package}

\item{v_i_list}{a list of fd functions v_i(t) :
\eqn{t_i = \int_0^T v_i(t) X(t) dt}}

\item{nb_comp}{a value, number of components to take into account, default NULL
if NULL, then delta(t) will take every components available.}
}
\value{
a fd function
}
\description{
This functions evaluates the regression function such as :
\eqn{Y = \int_0^T delta(t) X(t) dt}.
}
\author{
Francois Bassac
}
