% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{evaluate_results}
\alias{evaluate_results}
\title{evaluate_results
This function evaluates the PRESS, RMSE, MAE, R2 and the \% of variance between
Y and Y_hat}
\usage{
evaluate_results(Y, Y_hat)
}
\arguments{
\item{Y}{a vector of real values}

\item{Y_hat}{a vector of modeled values}
}
\value{
a dataframe
}
\description{
evaluate_results
This function evaluates the PRESS, RMSE, MAE, R2 and the \% of variance between
Y and Y_hat
}
\examples{
evaluate_results(c(1,2,3,4,5), c(0.9, 2.2, 4, 5.5, 5))

}
\author{
Francois Bassac
}
