% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{evaluate_V_i_function}
\alias{evaluate_V_i_function}
\title{evaluate_V_i_function}
\usage{
evaluate_V_i_function(w_i_list, gamma_ij_list)
}
\arguments{
\item{w_i_list}{a list of fd functions w_i(t)}

\item{gamma_ij_list}{a list of list of gamma_ij values}
}
\value{
a list of fd functions
}
\description{
This function evaluates the different functions v_i_t bases on gamma_ij and
the functions w_i_fd. Recursive :
\eqn{v_i = W_i - \sum_{j=1}^{i-1}gamma_ij*v_j}
}
\author{
Francois Bassac
}
