% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{evaluate_variance_explained}
\alias{evaluate_variance_explained}
\title{evaluate_variance_explained}
\usage{
evaluate_variance_explained(Y, Y_hat)
}
\arguments{
\item{Y}{a reference value}

\item{Y_hat}{a modelized value Y_hat = model(X)}
}
\value{
a value in \%
}
\description{
This function return the \% of variance explained by Y_hat comparing to Y.
}
\examples{
evaluate_variance_explained(c(1,2,3,4,5,6,7,8,9,10),
c(0.9, 1.1, 1.9, 2.4, 5.3, 6.01, 7,45, 9.12, 9.04, 11.6))

}
\author{
Francois Bassac
}
