% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{fit_pfr_multi}
\alias{fit_pfr_multi}
\title{Fit a Multivariate Penalized Functional Regression}
\usage{
fit_pfr_multi(D_global, Y, R_0_lambda)
}
\arguments{
\item{D_global}{The global design matrix (n x P_total)}

\item{Y}{The response vector (n x 1)}

\item{R_0_lambda}{The augmented global penalty matrix}
}
\value{
A list with the estimated coefficients, fitted values, and Hat diagonal
}
\description{
Fit a Multivariate Penalized Functional Regression
}
