% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{from_basis_to_fdlist}
\alias{from_basis_to_fdlist}
\title{from_basis_to_fdlist}
\usage{
from_basis_to_fdlist(basis)
}
\arguments{
\item{basis}{a basis fd object or a list of fd functions.}
}
\value{
a list of fd functions
}
\description{
This function transform if necessary the input basis into a list of fd
functions.
If basis is a basis object from fda, the output fd_list is the list of the
different basis functions as fd functions.
If basis is already a list of fd functions, nothing changes.
}
\examples{
basis = create_bspline_basis(start = 0, end = 10, nbasis = 5, norder = 4)
plot(basis)

basis_list = from_basis_to_fdlist(basis)

plot(basis_list[[1]], col = 1)
for(i in 2:length(basis_list)){
  plot(basis_list[[i]], add=TRUE, col = i)
}

basis_list_2 = from_basis_to_fdlist(basis_list)

}
\author{
Francois Bassac
}
