% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{from_fd_to_func}
\alias{from_fd_to_func}
\title{from_fd_to_func}
\usage{
from_fd_to_func(fd_obj = NULL, coef = NULL, basisobj = NULL)
}
\arguments{
\item{fd_obj}{a fd object to transform into a function}

\item{coef}{the coefficient of an fd object to transform into a function}

\item{basisobj}{the basis object of an fd object to transform into a function}
}
\value{
a function
}
\description{
This function transform a fd object into a function.
It require either the fd object OR the coefficient and the basis object.
}
\examples{
basis = create_bspline_basis(0, 100, 10, 4)
coef = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
func_from_fd = from_fd_to_func(coef = coef, basis = basis)

}
\author{
Francois Bassac
}
