% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpls.R
\name{funcPLS_predict}
\alias{funcPLS_predict}
\title{funcPLS_predict}
\usage{
funcPLS_predict(
  df_predict_list,
  delta_list,
  curve_type_obj = NULL,
  regul_time_obj = NULL,
  id_col_obj = "id",
  time_col_obj = "time",
  int_mode = 1,
  nb_pt = 10,
  subdivisions = 100,
  parallel = TRUE
)
}
\arguments{
\item{df_predict_list}{a list of dataframe (id, time, state_or_value)}

\item{delta_list}{a list of regression objects (Intercept, fd, etc)}

\item{curve_type_obj}{a list of the curves types 'cat' or 'num'}

\item{regul_time_obj}{a list of time regularization values}

\item{id_col_obj}{a list of characters of the names of the id columns}

\item{time_col_obj}{a list of characters of the names of the time columns}

\item{int_mode}{a integer for integration mode, 1 for integrate,
2 for pracma::trapz, default 1}

\item{nb_pt}{a integer, the number of intermediate points for integration
mode 2, default 10}

\item{subdivisions}{a integer, the number of subdivisions for integration
mode 1, default 100}

\item{parallel}{a boolean to use parallelization, default TRUE}
}
\value{
a numeric vector
}
\description{
This function performs the prediction on a df_predict_ms using the delta_list
\eqn{\hat(Y) = \delta_0 + \sum_{i=1}^K \int_0^T \delta_i(t) X_i(t) dt}.
}
\author{
Francois Bassac
}
