% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_probabilities}
\alias{generate_probabilities}
\title{generate_probabilities}
\usage{
generate_probabilities(N_proba)
}
\arguments{
\item{N_proba}{a int the number of values requested}
}
\value{
a vector of the probabilities
}
\description{
This function generates probabilities whose sum is 1.
}
\examples{
generate_probabilities(3)
generate_probabilities(5)

}
\author{
Francois Bassac
}
