% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_X_df}
\alias{generate_X_df}
\title{generate_X_df}
\usage{
generate_X_df(
  nind = 100,
  start = 0,
  end = 100,
  curve_type = "cat",
  lambda_0 = 0.2,
  lambda_1 = 0.1,
  prob_start = 0.5,
  noise_sd = 0.1,
  seed = 123
)
}
\arguments{
\item{nind}{number of individuals, default 100}

\item{start}{first time, default 0}

\item{end}{last time, default 100}

\item{curve_type}{character, type of wanted synthetic data, default 'cat', need to be \'cat\' or \'num\'.}

\item{lambda_0}{lambda parameter for exponential law for state 0, default 0.2}

\item{lambda_1}{lambda parameter for exponential law for state 1, default 0.1}

\item{prob_start}{Start state 1 probability, binomial law, default 0.5}

\item{noise_sd}{noise added to the signal, default 0.1}

\item{seed}{seed for reproducibility, default 123}
}
\value{
the dataframe of the individuals
}
\description{
This function generate synthetic data of nind X(t).
For 'cat' curve_type it is in state 0 or 1 by two
different exponential laws.
For curve_type = 'num' it is noised cosine.
}
\examples{
generate_X_df()

}
\author{
Francois Bassac
}
