% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_X_df_CFD}
\alias{generate_X_df_CFD}
\title{generate_X_df_CFD}
\usage{
generate_X_df_CFD(
  nind = 500,
  start = 0,
  end = 100,
  lambda_0 = 0.2,
  lambda_1 = 0.1,
  prob_start = 0.5,
  seed = 123
)
}
\arguments{
\item{nind}{number of individuals, default 500}

\item{start}{first time, default 0}

\item{end}{last time, default 100}

\item{lambda_0}{lambda parameter for exponential law for state 0, default 0.2}

\item{lambda_1}{lambda parameter for exponential law for state 1, default 0.1}

\item{prob_start}{Start state 1 probability, binomial law, default 0.5}

\item{seed}{a integer, random seed 123}
}
\value{
the dataframe of the individuals
}
\description{
This function generate synthetic data of nind X(t) in state 0 or 1 by two
different exponential laws.
}
\examples{
generate_X_df_CFD()
generate_X_df_CFD(10, 13, 60, 0.21, 0.13, 0.7)

}
\author{
Francois Bassac
}
