% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_X_df_multistates}
\alias{generate_X_df_multistates}
\title{generate_X_df_multistates}
\usage{
generate_X_df_multistates(
  nind = 100,
  N_states = 3,
  start = 0,
  end = 100,
  lambdas,
  transition_df,
  seed = 123
)
}
\arguments{
\item{nind}{a int of the number of the individuals, default 100}

\item{N_states}{a int of the number of states wanted, default 3}

\item{start}{a value of starting time, default 0}

\item{end}{a value of ending time, default 100}

\item{lambdas}{a vector of N_states lambda values from lambda_determination(N_states)}

\item{transition_df}{a dataframe with the transition matrix from tranfert_probabilities(N_states)}

\item{seed}{a integer, random seed}
}
\value{
a dataframe of a multistates Categorical Funcitonal Data
}
\description{
This function generates a multistates CFD.
}
\examples{
N_states = 4
lambdas = lambda_determination(N_states)
transition_df = transfert_probabilities(N_states)

df = generate_X_df_multistates(nind = 100, N_states, start=0, end=100,
lambdas,  transition_df)

}
\author{
Francois Bassac
}
