% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_X_df_SFD}
\alias{generate_X_df_SFD}
\title{generate_X_df_SFD}
\usage{
generate_X_df_SFD(nind = 100, start = 0, end = 100, noise_sd = 0.1, seed = 123)
}
\arguments{
\item{nind}{number of individuals, default 500}

\item{start}{first time, default 0}

\item{end}{last time, default 100}

\item{noise_sd}{noise added to the signal, default 0.1}

\item{seed}{seed for reproductability, default 123}
}
\value{
a dataframe
}
\description{
This function generate synthetic data of nind X(t) which are noised cosinuses.
}
\examples{
generate_X_df_SFD(nind = 100, start = 0, end = 100,
noise_sd = 0.1, seed = 123)

}
\author{
Francois Bassac
}
