% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_X_df_test}
\alias{generate_X_df_test}
\title{generate_X_df_test}
\usage{
generate_X_df_test(
  TTRatio = 0.2,
  nind = 500,
  start = 0,
  end = 100,
  curve_type = "cat",
  lambda_0 = 0.2,
  lambda_1 = 0.1,
  prob_start = 0.5,
  seed = 123
)
}
\arguments{
\item{TTRatio}{Train Test ratio, default 0.2}

\item{nind}{number of individuals, default 500}

\item{start}{first time, default 0}

\item{end}{last time, default 100}

\item{curve_type}{character, type of wanted synthetic data, default 'cat',
need to be \'cat\' or \'num\'.}

\item{lambda_0}{lambda parameter for exponential law for state 0, default 0.2}

\item{lambda_1}{lambda parameter for exponential law for state 1, default 0.1}

\item{prob_start}{Start state 1 probability, binomial law, default 0.5}

\item{seed}{a integer, random seed 123}
}
\value{
a dataframe of the test data
}
\description{
generate_X_df_test
}
\examples{
generate_X_df_test()
generate_X_df_test(TTRatio = 0.4, nind=8, start=0, end=10, curve_type = 'num')

}
\author{
Francois Bassac
}
