% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_Y_df_SFD}
\alias{generate_Y_df_SFD}
\title{generate_Y_df_SFD}
\usage{
generate_Y_df_SFD(
  df,
  beta_real_func,
  beta_0_real = 5.4321,
  NotS_ratio = 0.2,
  id_col = "id",
  time_col = "time",
  seed = 123
)
}
\arguments{
\item{df}{the X(t) dataframe to evaluate Y on}

\item{beta_real_func}{a function beta(t), function used for the Y evaluation}

\item{beta_0_real}{the intercept, default 5.4321}

\item{NotS_ratio}{the Noise over total Signal ratio, default 0.2}

\item{id_col}{a character of the id column, default 'id'}

\item{time_col}{a character of the time column, default 'time'}

\item{seed}{a integer, random seed, 123}
}
\value{
a dataframe
}
\description{
This function generates Y_df bases on df, beta_func with the following link
Y = beta_0 + int(X(t)*beta(t))dt
It generates also the noised values of Y.
Here the NotS_ratio is the Noise over total Signal ratio meaning that a
value of 0.2 means that the noise represents 20\% of the TOTAL variance.
}
\author{
Francois Bassac
}
