% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{get_auto_candidate_list}
\alias{get_auto_candidate_list}
\title{Generate an automatic smart lambda grid based on trace heuristics}
\usage{
get_auto_candidate_list(block_list, block_sizes, length_out = 5)
}
\arguments{
\item{block_list}{A list of blocks generated by the build_block_* functions.}

\item{block_sizes}{A numeric vector indicating how many parameters belong to each block.}

\item{length_out}{Integer. The number of candidate values per lambda parameter.}
}
\value{
A flat list of candidate vectors, ready for generate_lambda_grid.
}
\description{
Generate an automatic smart lambda grid based on trace heuristics
}
