% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{gram_schmidt_orthonormalize}
\alias{gram_schmidt_orthonormalize}
\title{gram_schmidt_orthonormalize}
\usage{
gram_schmidt_orthonormalize(basis, output_type = "fdlist", tol = 1e-12)
}
\arguments{
\item{basis}{A basis object from fda package or a list of fd functions.}

\item{output_type}{A character to choose the output format. "fdlist"
(default) or "funlist" for R functions.}

\item{tol}{a float, tolerance parameter, default 1e-12}
}
\value{
A list of orthonormalized functions fd or func(t)
}
\description{
Orthonormalize a basis functions with Gram-Schmidt algorithm.
}
\examples{

start = 0
end = 10
basis = create_bspline_basis(start, end, nbasis = 10, norder = 4)

basis_orth = gram_schmidt_orthonormalize(basis, "fdlist")

}
\author{
Francois Bassac
}
