% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{help_smoothPLS}
\alias{help_smoothPLS}
\title{help_smoothPLS}
\usage{
help_smoothPLS()
}
\value{
No return value, called for side effects (prints instructions to the console).
}
\description{
This function print some information on the package use.
}
\author{
Francois Bassac
}
