% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{initial_state_determination}
\alias{initial_state_determination}
\title{initial_state_determination}
\usage{
initial_state_determination(N_states)
}
\arguments{
\item{N_states}{a int the number of considered states}
}
\value{
a value of the designated state
}
\description{
This functions determine randomly the initial state of a categorical
functional data, uniform probability between states.
}
\examples{
initial_state_determination(2)
initial_state_determination(7)

}
\author{
Francois Bassac
}
