% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{is_orthogonal}
\alias{is_orthogonal}
\title{is_orthogonal}
\usage{
is_orthogonal(basis, tol = 1e-10)
}
\arguments{
\item{basis}{A basis object from fda package or a list of fd functions.}

\item{tol}{a float, tolerance parameter, default 1e-10)}
}
\value{
A boolean, TRUE if orthogonal, FALSE if not
}
\description{
Check if a basis function is orthogonal
}
\author{
Francois Bassac
}
