% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{lambda_augmented_uni}
\alias{lambda_augmented_uni}
\title{Augment a lambda matrix with an intercept column}
\usage{
lambda_augmented_uni(lambda)
}
\arguments{
\item{lambda}{A numeric matrix or data frame containing the lambda values.}
}
\value{
A matrix with an additional first column named "intercept"
filled with ones.
}
\description{
Prepends a column of ones to a given matrix, effectively adding an
intercept term for regression models.
}
\examples{
lam <- matrix(1:4, nrow = 2)
lambda_augmented_uni(lam)
}
\author{
Francois Bassac
}
