% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{lambda_determination}
\alias{lambda_determination}
\title{lambda_determination}
\usage{
lambda_determination(N_states, lambda_values = c(0.05, 0.25))
}
\arguments{
\item{N_states}{a int the number of states}

\item{lambda_values}{a vector of min and max values authorized for lambda}
}
\value{
a numeric vector with the lambda values
}
\description{
This function determines a number of lambda parameters for exponential laws.
}
\examples{
lambda_determination(3)
lambda_determination(7)

}
\author{
Francois Bassac
}
