% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpls.R
\name{multivariate_alpha_building}
\alias{multivariate_alpha_building}
\title{multivariate_alpha_building}
\usage{
multivariate_alpha_building(
  df_list,
  basis_list,
  curve_type_list,
  regul_time_list,
  id_col_list,
  time_col_list,
  print_steps = FALSE
)
}
\arguments{
\item{df_list}{a list of data frame (id, time, state_or_value)}

\item{basis_list}{a list of basis}

\item{curve_type_list}{a list of curves}

\item{regul_time_list}{a list of regul_time}

\item{id_col_list}{a list if the characters for the id column}

\item{time_col_list}{a list of the characters for the time column}

\item{print_steps}{a boolean to print the current step, default FALSE}
}
\value{
a list of the alphas matrix and the curve_names vector
and the new_basis_list
}
\description{
This function builds the alphas matrix by columns binding the alpha of each
curve.
It returne the alphas matrix and a vector containing all the curves names.
}
\author{
Francois Bassac
}
