% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpls.R
\name{multivariate_assemble_basis_metric}
\alias{multivariate_assemble_basis_metric}
\title{multivariate_assemble_basis_metric}
\usage{
multivariate_assemble_basis_metric(
  df_list,
  basis_list,
  curve_list,
  id_col_list,
  time_col_list
)
}
\arguments{
\item{df_list}{a list of dataframe (id, time, value_or_state)}

\item{basis_list}{a list of basis fd object}

\item{curve_list}{a list of the curves type 'cat' or 'num'}

\item{id_col_list}{a list of the character of the id columns}

\item{time_col_list}{a list of the character of the time columns}
}
\value{
a matrix of the metric to consider
}
\description{
This function assemble the metrics of all the basis of the basis list for
multivariate use case.
It take care on the categorical case by de-multiplying the basis per state.
}
\author{
Francois Bassac
}
