% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_pls.R
\name{naivePLS}
\alias{naivePLS}
\title{naivePLS}
\usage{
naivePLS(
  df_list,
  Y,
  regul_time_obj = NULL,
  curve_type_obj = NULL,
  id_col_obj = "id",
  time_col_obj = "time",
  print_steps = FALSE,
  plot_rmsep = TRUE,
  print_nbComp = TRUE,
  plot_reg_curves = FALSE,
  validation = "LOO",
  jackknife = TRUE
)
}
\arguments{
\item{df_list}{a list of dataframe (id, time, value_or_state)}

\item{Y}{a numeric vector for the scalar response}

\item{regul_time_obj}{a list of time regularization values}

\item{curve_type_obj}{a list of the curve types 'cat' or 'num'}

\item{id_col_obj}{a list of character of the names of the id columns}

\item{time_col_obj}{a list of character of the names of the time columns}

\item{print_steps}{a boolean to print the different steps, default FALSE}

\item{plot_rmsep}{a boolean to plot the RMSEP, default TRUE}

\item{print_nbComp}{a boolean to print the optimal number or components,
default TRUE}

\item{plot_reg_curves}{a boolean to plot the regression curves, default
FALSE}

\item{validation}{a character, pls::plsr input, default 'LOO'}

\item{jackknife}{a boolean, pls::plsr input, default TRUE}
}
\value{
a list of ("plsr_model", "nbCP_opti", "curves_names",
"opti_reg_coef", "reg_obj")
}
\description{
This function performs the naive PLS method for Categorical functional data,
Scalar functional data and multivariate data.
}
\author{
Francois Bassac
}
