% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_pls.R
\name{naivePLS_formating}
\alias{naivePLS_formating}
\title{naivePLS_formating}
\usage{
naivePLS_formating(
  df_list,
  regul_time_obj = NULL,
  curve_type_obj = NULL,
  id_col_obj = "id",
  time_col_obj = "time"
)
}
\arguments{
\item{df_list}{a list of dataframeaof 3 columns (id, time, state_or_value)}

\item{regul_time_obj}{a list of vector of time regularization values}

\item{curve_type_obj}{a list of character of the type of curves}

\item{id_col_obj}{a list of character for the id column names}

\item{time_col_obj}{a list of character for the time column names}
}
\value{
a list
}
\description{
This function format the list of given dataframe into a time regularized
matrix usable for the naivePLS function or prediction.
}
\author{
Francois Bassacs
}
