% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_pls.R
\name{naivePLS_predict}
\alias{naivePLS_predict}
\title{naivePLS_predict}
\usage{
naivePLS_predict(
  naive_pls_obj,
  df_predict_list,
  regul_time_obj = NULL,
  curve_type_obj = NULL,
  id_col_obj = "id",
  time_col_obj = "time"
)
}
\arguments{
\item{naive_pls_obj}{a list of a naivePLS object}

\item{df_predict_list}{a list of dataframe (id, time, state_or_value)}

\item{regul_time_obj}{a list of time regularization values}

\item{curve_type_obj}{a list of curves types 'cat' or 'num'}

\item{id_col_obj}{a list of id column names}

\item{time_col_obj}{a list of time column names}
}
\value{
a numeric vector
}
\description{
This function use the df_pred to make a prediction for a naivePLS object.
}
\author{
Francois Bassac
}
