% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{number_of_test_id}
\alias{number_of_test_id}
\title{number_of_test_id}
\usage{
number_of_test_id(TTRatio = 0.2, nind = 100)
}
\arguments{
\item{TTRatio}{Train Test ratio, default 0.2}

\item{nind}{number of individuals, default 100}
}
\value{
int, the number of id for test set
}
\description{
This function evaluates the number of individuals for a test set.
}
\examples{
number_of_test_id(TTRatio = 0.2, nind=100)

}
\author{
Francois Bassac
}
