% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{obj_list_creation}
\alias{obj_list_creation}
\title{obj_list_creation}
\usage{
obj_list_creation(N_rep, obj)
}
\arguments{
\item{N_rep}{a value of the number of states sharing the same basis}

\item{obj}{a object}
}
\value{
a list of N_rep appended obj
}
\description{
This functions creates a list of basis containing the
value of the number of states or curves sharing the same basis.
}
\examples{
N_rep = 4
start = 1
end = 51
nbasis = 13
norder = 3

basis = create_bspline_basis(start, end, nbasis, norder)
basis_list = obj_list_creation(N_rep, basis)
obj_list_creation(N_rep, 0:100)

}
\author{
Francois Bassac
}
