% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{orthonormalize_basis_list}
\alias{orthonormalize_basis_list}
\title{orthonormalize_basis_list}
\usage{
orthonormalize_basis_list(basis_list, orth_list, tol = 1e-09)
}
\arguments{
\item{basis_list}{a list of basis to orthonormalized}

\item{orth_list}{a list of boolean per basis if orthonormalization is needed}

\item{tol}{a float, tolerance parameter.}
}
\value{
a list of list of fd object : the orthonormalized functions.
}
\description{
This function orthonormalized a basis from a list if needed.
}
\author{
Francois Bassac
}
