% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{p_w_building}
\alias{p_w_building}
\title{p_w_building}
\usage{
p_w_building(
  coefficient,
  N_curves_processed,
  new_basis_list,
  new_orth_basis_list,
  curves_names_list
)
}
\arguments{
\item{coefficient}{a table of the coefficient to use}

\item{N_curves_processed}{a integer, the real number of curves
(1 sfd = 1 curve, 1 cfd = 1curve per state)}

\item{new_basis_list}{a list of processed basis list
(1 basis per curve_processed)}

\item{new_orth_basis_list}{a list of orthonormalized basis list
(1 basis per curve_processed)}

\item{curves_names_list}{a list of curve name (1 name per curve_processed)}
}
\value{
a list of fd objects
}
\description{
This functions evaluates the p_i(t) (X_i regression functions) and w_i(t)
such as \eqn{t_i = \int_0^T w_i(t)X_{i-1}(t) dt}.
The evaluation is done for all the components.
}
\author{
Francois Bassac
}
