% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{plot.cv_pfr_uni}
\alias{plot.cv_pfr_uni}
\title{Plot method for cv_pfr_uni objects}
\usage{
\method{plot}{cv_pfr_uni}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cv_pfr_uni}.}

\item{...}{Additional graphical parameters.}
}
\description{
Visualizes the LOOCV RMSEP curve as a function of the penalty parameter lambda.
}
\author{
Francois Bassac
}
