% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{plot_fd_list}
\alias{plot_fd_list}
\title{plot_fd_list}
\usage{
plot_fd_list(fd_list, curves_names, regul_time)
}
\arguments{
\item{fd_list}{a list of fd objects}

\item{curves_names}{a list of the curves names}

\item{regul_time}{a numeric vector}
}
\value{
a plot
}
\description{
This function plots on the same figure the fd curves from the fd_list by
evaluating them on the given regul_time
}
\author{
Francois Bassac
}
