% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{plot_real_and_smoothed_data_ind}
\alias{plot_real_and_smoothed_data_ind}
\title{plot_real_and_smoothed_data_ind}
\usage{
plot_real_and_smoothed_data_ind(
  df_wide,
  df_fd,
  time_seq = 0:100,
  id = 1,
  col_list = c("blue", "red", "green", "yellow")
)
}
\arguments{
\item{df_wide}{a wide dataframe, output of convert_to_wide_format()}

\item{df_fd}{a list of functional data from df_wide}

\item{time_seq}{a vector to plot on, default 0:100}

\item{id}{a value of the id to plot}

\item{col_list}{a list of color to separate the states}
}
\value{
a plot
}
\description{
plot_real_and_smoothed_data_ind
}
\examples{
N_states = 3
lambdas = lambda_determination(N_states)
transition_df = transfer_probabilities(N_states)

df = generate_X_df_multistates(nind = 100, N_states, start=0, end=100,
lambdas,  transition_df)
df_processed = cat_data_to_indicator(df)

df_regul = list()
df_wide = list()
for(name in names(df_processed)){
print(paste0(name, " regularisation"))
df_regul[[name]] = regularize_time_series(df_processed[[name]],
time_seq =  c(0:100), curve_type = 'cat', id_col='id', time_col='time')

df_wide[[name]] = convert_to_wide_format(df_regul[[name]], id_col='id',
time_col='time')
}

basis = create_bspline_basis(0, 100, 10, 4)

df_fd = list()
for(name in names(df_wide)){
print(paste0(name, " fd transformation"))
df_fd[[name]] = fda::Data2fd(argvals = c(0:100),
y = t(df_wide[[name]][, -c(1)]), basis)
}

plot_real_and_smoothed_data_ind(df_wide, df_fd, c(0:100), id=1)

}
\author{
Francois Bassac
}
