% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_functional_regression.R
\name{preprocess_mpfr_data}
\alias{preprocess_mpfr_data}
\title{Preprocess data list for MPFR}
\usage{
preprocess_mpfr_data(data_list, types, id_col = "id", time_col = "time")
}
\arguments{
\item{data_list}{The raw list of predictors.}

\item{types}{The types of the predictors.}

\item{id_col}{The name of the ID column in functional dataframes.}

\item{time_col}{The name of the time column in functional dataframes.}
}
\value{
A standardized list of data objects.
}
\description{
Preprocess data list for MPFR
}
