% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{press_model}
\alias{press_model}
\title{press_model
This function evaluates the PRESS error 'LOO' of a lm or glm model.}
\usage{
press_model(model)
}
\arguments{
\item{model}{a model from lm or glm}
}
\value{
a value
}
\description{
press_model
This function evaluates the PRESS error 'LOO' of a lm or glm model.
}
\author{
Francois Bassac
}
