% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{press_values}
\alias{press_values}
\title{press_values}
\usage{
press_values(y, y_hat)
}
\arguments{
\item{y}{a vector of real values}

\item{y_hat}{a vector of predicted values}
}
\value{
a value
}
\description{
This function evaluates the press error using values y and y_hat
}
\examples{
y = c(1, 2, 4, 6, 8, 10)
y_hat = c(2, 3, 5, 7, 9, 11)
press_values(y, y_hat)

}
\author{
Francois Bassac
}
