% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{r_squared_values}
\alias{r_squared_values}
\title{r_squared_values}
\usage{
r_squared_values(y, y_hat)
}
\arguments{
\item{y}{a vector of real values}

\item{y_hat}{a vector of predicted values}
}
\value{
a value
}
\description{
This function evaluates the R_2 using values y and y_hat
}
\examples{
y = c(1, 2, 4, 6, 8, 10)
y_hat = c(2, 3, 5, 7, 9, 11)
r_squared_values(y, y_hat)

}
\author{
Francois Bassac
}
