% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpls.R
\name{reg_curve_funcPLS_evaluation}
\alias{reg_curve_funcPLS_evaluation}
\title{reg_curve_funcPLS_evaluation}
\usage{
reg_curve_funcPLS_evaluation(
  mfpls_mfd,
  nb_comp_pls_opt,
  root_metric,
  new_basis_list,
  curve_names,
  print_steps = FALSE
)
}
\arguments{
\item{mfpls_mfd}{a MFPLS model}

\item{nb_comp_pls_opt}{the optimal number of components}

\item{root_metric}{the root metric}

\item{new_basis_list}{a list of basis fd objects}

\item{curve_names}{a list of curves to keep}

\item{print_steps}{a boolean to cat or not the steps}
}
\value{
a list of the betas per state
}
\description{
This function evaluate the beta regression functions of the multivariate
functional PLS.
}
\author{
Francois Bassac
}
