% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{regularize_time_series}
\alias{regularize_time_series}
\title{regularize_time_series}
\usage{
regularize_time_series(
  df,
  time_seq = 0:100,
  curve_type = NULL,
  id_col = "id",
  time_col = "time"
)
}
\arguments{
\item{df}{dataframe with one or more different ids}

\item{time_seq}{New time sequence where we want to regularize}

\item{curve_type}{A string giving the type of the curve, 'cat' for a
categorical functional data, 'num' for a scalar functional data, default : NULL}

\item{id_col}{col_name of df for the id}

\item{time_col}{col_name of df for the time}
}
\value{
a dataframe with the regularized data on time_seq
}
\description{
This function regularize the data on a new time interval time_seq given
the curve_type.
For curve_type = 'cat' the output state stay the same between 2 times.
For curve_type = 'num' the intermediate values are interpolated linearly.
}
\examples{
id_df = data.frame(id=rep(1,5), time=seq(0, 40, 10), state=c(0, 1, 1, 0, 1))
regularize_time_series(id_df, time_seq = seq(0, 40, 2), curve_type = 'cat')
regularize_time_series(id_df, time_seq = seq(0, 40, 2), curve_type = 'num')

}
\author{
Francois Bassac
}
