% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{regularize_time_series_CFD}
\alias{regularize_time_series_CFD}
\title{regularize_time_series_CFD}
\usage{
regularize_time_series_CFD(
  df,
  time_seq = 0:100,
  id_col = "id",
  time_col = "time"
)
}
\arguments{
\item{df}{dataframe with one or more different ids}

\item{time_seq}{New time sequence where we want to regularize}

\item{id_col}{col_name of df for the id}

\item{time_col}{col_name of df for the time}
}
\value{
a dataframe with the regularized data on time_seq
}
\description{
This function regularize the data CFD on a new time interval time_seq
}
\author{
Francois Bassac
}
