% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{regularize_time_series_SFD}
\alias{regularize_time_series_SFD}
\title{regularize_time_series_SFD}
\usage{
regularize_time_series_SFD(
  df,
  time_seq = 0:100,
  id_col = "id",
  time_col = "time"
)
}
\arguments{
\item{df}{dataframe with one or more different ids}

\item{time_seq}{New time sequence where we want to regularize}

\item{id_col}{col_name of df for the id}

\item{time_col}{col_name of df for the time}
}
\value{
a regularized dataframe
}
\description{
This function regularizz a Scalar Functional Data SFD on the time_seq input.
This function uses linear interpolation.
}
\author{
Francois Bassac
}
