% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{remove_duplicate_states}
\alias{remove_duplicate_states}
\title{remove_duplicate_states}
\usage{
remove_duplicate_states(data, id_col = "id", time_col = "time")
}
\arguments{
\item{data}{a single- or multi-state dataframe.}

\item{id_col}{a character for the id column, default 'id'}

\item{time_col}{a character for the time column, default 'time'}
}
\value{
a dataframe with no duplicated states.
}
\description{
This function removes the duplicated states and keep the last line.
this function works both with (0,1) or categorical states on the state_col!
}
\examples{
N_states = 3
lambdas = lambda_determination(N_states)
transition_df = transfert_probabilities(N_states)

df = generate_X_df_multistates(nind = 100, N_states, start=0, end=100,
lambdas, transition_df)
remove_duplicate_states(df)

}
\author{
Francois Bassac
}
