% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{select_from_fd_list}
\alias{select_from_fd_list}
\title{select_from_fd_list}
\usage{
select_from_fd_list(fd_list, toDrop = NULL)
}
\arguments{
\item{fd_list}{a list of fd object}

\item{toDrop}{a numeric vector}
}
\value{
a list of fd objects
}
\description{
This function return a list of functions of fd_list without the ones
specified in the numeric vector toDrop.
}
\author{
Francois Bassac
}
