% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{smoothPLS_CFD_predict_para_v1}
\alias{smoothPLS_CFD_predict_para_v1}
\title{smoothPLS_CFD_predict_para_v1 (Updated v0.1.2)}
\usage{
smoothPLS_CFD_predict_para_v1(
  df_predict,
  delta_spls,
  id_col = "id",
  time_col = "time",
  subdivisions = 100,
  parallel = TRUE,
  ...
)
}
\arguments{
\item{df_predict}{Dataframe containing columns for id, time, and state.}

\item{delta_spls}{A list containing the scalar intercept and the functional
regression coefficient (fd object).}

\item{id_col}{Character, name of the id column, default 'id'.}

\item{time_col}{Character, name of the time column, default 'time'.}

\item{subdivisions}{integer, maximum number of sub-intervals for integration,
default 100}

\item{parallel}{a boolean to enable parallel processing, default TRUE.}

\item{...}{Additional parameters passed to evaluate_id_func_integral
(e.g., rel_tol, subdivisions).}
}
\value{
A numeric vector of predicted values for each individual.
}
\description{
Predicts the response variable for Categorical Functional Data (CFD) by
integrating the regression coefficient function over active state intervals.
}
\author{
Francois Bassac
}
