% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{smoothPLS_predict}
\alias{smoothPLS_predict}
\title{smoothPLS_multi_predict}
\usage{
smoothPLS_predict(
  df_predict_list,
  delta_list,
  curve_type_obj = NULL,
  id_col_obj = "id",
  time_col_obj = "time",
  regul_time_obj = NULL,
  int_mode = 1,
  nb_pt = 10,
  subdivisions = 100,
  parallel = TRUE
)
}
\arguments{
\item{df_predict_list}{a list of dataframe (id, time, value_or_state)}

\item{delta_list}{a list of regression object (intercept, delta_1_fd,
delta_2_fd, etc)}

\item{curve_type_obj}{a list of characters of the curve types 'cat' or 'num'}

\item{id_col_obj}{a list of character of the name of the id column,
default 'id'}

\item{time_col_obj}{a list of character of the name of the time column,
default 'time'}

\item{regul_time_obj}{a list of the time regularization values}

\item{int_mode}{a integer for the integration mode, 1 for integrate, 2 for
pracma::trapz}

\item{nb_pt}{a integrer, number of intermediate points for pracma::trapz,
default 10}

\item{subdivisions}{a integer, number of subdivision in integrate function,
default 100}

\item{parallel}{a boolean to use parallelization, default TRUE}
}
\value{
a numeric vector of the prediction
}
\description{
This function use the list of regression functions to make a prediction
\eqn{\hat{Y} = \delta_0 + \int_0^T \delta(t) X(t) dt}
}
\author{
Francois Bassac
}
