% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{smoothPLS_predict_uni}
\alias{smoothPLS_predict_uni}
\title{smoothPLS_predict_uni}
\usage{
smoothPLS_predict_uni(
  df_predict,
  delta_list,
  curve_type = NULL,
  int_mode = 1,
  id_col = "id",
  time_col = "time",
  nb_pt = 10,
  subdivisions = 100,
  regul_time = seq(delta_list[[2]]$basis$rangeval[1], delta_list[[2]]$rangeval[2], 1)
)
}
\arguments{
\item{df_predict}{a dataframe ('id', 'time', 'state or value') to predict
from}

\item{delta_list}{a list of delta_spls : list(intercept, delta_fd)}

\item{curve_type}{a character, 'cat' for Categorical FD, 'num' for Scalar FD}

\item{int_mode}{a value of the integration mode, default 1}

\item{id_col}{a character for the id column, default 'id'}

\item{time_col}{a character for the time column, default 'time'}

\item{nb_pt}{number of points for the integration, default value : 10}

\item{subdivisions}{default parameter of R function integrate;
default value : 100}

\item{regul_time}{a vector of time regularization values default delta_fd
basis rangeval per 1, NEEDED for curve_type = 'num'!}
}
\value{
a vector of predicted values
\eqn{\hat{Y} = \delta_0 + \int_0^T X(t) \delta(t) dt}
}
\description{
This function make a prediction base on a dataframe and a list made of the
intercept and the regression curve. The input curve_type in needed to select
the good way of evaluate the integrals \eqn{\int_0^T delta(t) X(t) dt}.
}
\author{
Francois Bassac
}
