% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{smoothPLS_SFD_predict}
\alias{smoothPLS_SFD_predict}
\title{smoothPLS_SFD_predict}
\usage{
smoothPLS_SFD_predict(
  df_predict,
  delta_spls,
  regul_time = seq(delta_spls[[2]]$basis$rangeval[1], delta_spls[[2]]$basis$rangeval[2],
    1),
  int_mode = 2,
  id_col = "id",
  time_col = "time",
  nb_pt = 10,
  subdivisions = 100
)
}
\arguments{
\item{df_predict}{a dataframe ('id', 'time', 'state') to predict from}

\item{delta_spls}{a list of delta_spls}

\item{regul_time}{a vector of time regularization values default delta_fd
basis rangeval per 1}

\item{int_mode}{a value of the integration mode, default 1}

\item{id_col}{a character for the id column, default 'id'}

\item{time_col}{a character for the time column, default 'time'}

\item{nb_pt}{number of points for the integration, default value : 10}

\item{subdivisions}{default parameter of R function integrate;
default value : 100}
}
\value{
a vector of predicted values
}
\description{
This function make prediction on new data from pls-cfd delta
Y = delta_0 + int delta_1(t) X(t) dt.
This function can be used for a classic functional PLS prediction.
}
\author{
Francois Bassac
}
