% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{test_basis_properties}
\alias{test_basis_properties}
\title{test_basis_properties}
\usage{
test_basis_properties(basis, name = "Base", tol = 1e-10)
}
\arguments{
\item{basis}{Basis to test, basis object or list of fd functions}

\item{name}{Character, name of the basis}

\item{tol}{Float, precision, default 1e-10}
}
\value{
a boolean
}
\description{
General function to test basis functions
}
\examples{
start = 0
end = 10
basis = create_bspline_basis(start, end, nbasis=10, norder=4)
test_basis_properties(basis, "cubic splines", tol = 1e-3)

basis_f = fda::create.fourier.basis(rangeval=c(start, end), nbasis=5)
test_basis_properties(basis_f, "Fourier", tol = 1e-3)

}
\author{
Francois Bassac
}
