% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{transfer_probabilities}
\alias{transfer_probabilities}
\title{transfer_probabilities}
\usage{
transfer_probabilities(N_states)
}
\arguments{
\item{N_states}{a int the number a states considered.}
}
\value{
a dataframe containing the transition probabilities
}
\description{
This function gives transfer probabilities between states. row -> columns.
}
\examples{
transfer_probabilities(3)
transfer_probabilities(5)

}
\author{
Francois Bassac
}
