% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{transition_matrix}
\alias{transition_matrix}
\title{transition_matrix}
\usage{
transition_matrix(basis1, basis2)
}
\arguments{
\item{basis1}{First basis, basis obj or list of fd functions}

\item{basis2}{Second basis, basis obj or list of fd functions}
}
\value{
Transition matrix P such as basis1 = P * basis2
}
\description{
Build transition matrix between to basis.
}
\author{
Francois Bassac
}
