% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpls.R
\name{univariate_alpha_building}
\alias{univariate_alpha_building}
\title{univariate_alpha_building}
\usage{
univariate_alpha_building(
  df,
  basis,
  curve_type = NULL,
  regul_time,
  id_col = "id",
  time_col = "time"
)
}
\arguments{
\item{df}{a dataframe (id, time, value_or_state)}

\item{basis}{a basis fd object}

\item{curve_type}{a character, 'cat' or 'num'}

\item{regul_time}{a numeric vector for time regularization}

\item{id_col}{a character for the id column name}

\item{time_col}{a character for the time column name}
}
\value{
a matrix
}
\description{
This function build the alphas matrix for a dataframe of a curve_type curve
on a basis on a regul_time vector.
}
\author{
Francois Bassac
}
