% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{plot_model_metrics_base}
\alias{plot_model_metrics_base}
\title{plot_model_metrics_base}
\usage{
plot_model_metrics_base(
  train_results,
  test_results,
  models_to_plot = c("FPLS", "SmoothPLS", "NaivePLS"),
  n_digits = 3
)
}
\arguments{
\item{train_results}{a dataframe of train results}

\item{test_results}{a dataframe of test results}

\item{models_to_plot}{a list of characters of the models to plot, default
c("FPLS", "SmoothPLS", "NaivePLS")}

\item{n_digits}{a integer for the number of significative numbers to print,
default 3}
}
\value{
a plot
}
\description{
This funciton plots some histograms for train_results and test_results
}
\author{
Francois Bassac
}
