% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_PLS.R
\name{smoothPLS_SFD_predict}
\alias{smoothPLS_SFD_predict}
\title{smoothPLS_SFD_predict}
\usage{
smoothPLS_SFD_predict(
  df_predict,
  delta_spls,
  basis_obj = NULL,
  id_col = "id",
  time_col = "time",
  ...
)
}
\arguments{
\item{df_predict}{Dataframe with columns (id, time, value).}

\item{delta_spls}{List containing (intercept, delta_fd_object).}

\item{basis_obj}{Optional basis for signal reconstruction. If NULL, uses the basis from delta_fd}

\item{id_col}{Character, name of id column.}

\item{time_col}{Character, name of time column.}

\item{...}{Additional arguments for Data2fd or inprod.}
}
\value{
A numeric vector of predicted values
}
\description{
Predicts the response Y for Scalar Functional Data using the analytic L2 inner product.
This implementation follows the theory from Chapter 7.
}
